Fix -[UIImage(TTCategory) transformWidth:height:rotate:] "invalid context"
This method is meant to resize an image, but it was using the wrong value for
the bytesPerRow of the new image (it wasn't taking into account the fact that
the image is being resized -- facepalm). The bug would only manifest if the
destination size was bigger than the image's current size, because the value
given for the new image's bytesPerRow would be too small. If the image was being
scaled down, it would work properly, but some memory would be wasted because the
new image's bytesPerRow would be bigger than necessary. Sigh.

I also wrote a brief doc comment for the method while I was at it.