Use DidShow and DidHide keyboard notifications instead of custom hacks
There are actually official notifications of when the keyboard is finished
hiding and showing, so use those instead of using timers with guessed
delays. Also, use the symbols for the notification names instead of strings.

This fixes the bug wherein the keyboard events arrive in a weird order when the
first responder changes from one control to another immediately. Previously it
went will-disappear, will-appear, did-disappear, did-appear, which doesn't make
sense. Now, it goes will-disappear, will-appear, did-appear -- i.e. it doesn't
lie and say the keyboard disappeared fully, and the events arrive in a sensible
order.

There is still some weird behavior specific to table view controllers
(willDisappear sets the frame and then it immediately gets reset in -didAppear,
causing some weird jumping) but that's for a separate change.